requirements.txt
# Core RAG Framework
haystack-ai>=2.0.0

# LLM Integration
anthropic-haystack

# Embeddings
sentence-transformers  # For generating text embeddings

# Vector Database
chromadb  # For persistent local or Docker deployment
chroma-haystack  # Haystack integration for Chroma

# Alternative Vector Databases (currently not used)
# faiss-cpu  # For local development
# qdrant-client  # For Qdrant vector database (TODO: Consider for production)
# weaviate-client  # For Weaviate

# Web Crawling
crawlee  # Modern web crawling framework
playwright  # Browser automation for JS-heavy sites
html2text  # Convert HTML to markdown
readability-lxml  # Extract main content from HTML

# Document Processing
pypdf  # PDF support
python-docx  # Word document support
markdown  # Markdown parsing
beautifulsoup4  # HTML parsing
lxml  # XML/HTML processing

# Utilities
python-dotenv  # Environment variable management
pydantic  # Data validation
pyyaml  # Configuration files

# API/Interface (if building REST API)
fastapi  # Web framework
uvicorn  # ASGI server

# Development
pytest  # Testing
black  # Code formatting
ruff  # Linting
